+++ driver/src/main/java/oracle/nosql/driver/http/Client.java
@@ -15,11 +15,14 @@ import static oracle.nosql.driver.util.BinaryProtocol.DEFAULT_SERIAL_VERSION;
 import static oracle.nosql.driver.util.BinaryProtocol.V2;
 import static oracle.nosql.driver.util.BinaryProtocol.V3;
 import static oracle.nosql.driver.util.CheckNull.requireNonNull;
+<<<<<<< HEAD
 import static oracle.nosql.driver.util.LogUtil.isLoggable;
 import static oracle.nosql.driver.util.LogUtil.logFine;
 import static oracle.nosql.driver.util.LogUtil.logInfo;
 import static oracle.nosql.driver.util.LogUtil.logTrace;
 import static oracle.nosql.driver.util.LogUtil.logWarning;
+=======
+>>>>>>> 752c1eaadcac8dce3f74edb9a1e243701f31cb04
 import static oracle.nosql.driver.util.HttpConstants.ACCEPT;
 import static oracle.nosql.driver.util.HttpConstants.CONNECTION;
 import static oracle.nosql.driver.util.HttpConstants.CONTENT_LENGTH;
@@ -27,6 +30,10 @@ import static oracle.nosql.driver.util.HttpConstants.CONTENT_TYPE;
 import static oracle.nosql.driver.util.HttpConstants.NOSQL_DATA_PATH;
 import static oracle.nosql.driver.util.HttpConstants.REQUEST_ID_HEADER;
 import static oracle.nosql.driver.util.HttpConstants.USER_AGENT;
+import static oracle.nosql.driver.util.LogUtil.isLoggable;
+import static oracle.nosql.driver.util.LogUtil.logFine;
+import static oracle.nosql.driver.util.LogUtil.logInfo;
+import static oracle.nosql.driver.util.LogUtil.logTrace;
 
 import java.io.IOException;
 import java.net.URL;
@@ -34,8 +41,8 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Executors;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -55,6 +62,7 @@ import oracle.nosql.driver.RequestTimeoutException;
 import oracle.nosql.driver.RetryHandler;
 import oracle.nosql.driver.RetryableException;
 import oracle.nosql.driver.SecurityInfoNotReadyException;
+import oracle.nosql.driver.StatsControl;
 import oracle.nosql.driver.WriteThrottlingException;
 import oracle.nosql.driver.UnsupportedProtocolException;
 import oracle.nosql.driver.httpclient.HttpClient;
@@ -161,10 +169,17 @@ public class Client {
      */
     private ExecutorService threadPool;
 
+<<<<<<< HEAD
     private short serialVersion = DEFAULT_SERIAL_VERSION;
 
     /* for one-time messages */
     private final HashSet<String> oneTimeMessages;
+=======
+    /**
+     * config for statistics
+     */
+    private StatsControlImpl statsControl;
+>>>>>>> 752c1eaadcac8dce3f74edb9a1e243701f31cb04
 
     public Client(Logger logger,
                   NoSQLHandleConfig httpConfig) {
@@ -235,7 +250,12 @@ public class Client {
             threadPool = null;
         }
 
+<<<<<<< HEAD
         oneTimeMessages = new HashSet<String>();
+=======
+        statsControl = new StatsControlImpl(config,
+            logger, httpClient, httpConfig.getRateLimitingEnabled());
+>>>>>>> 752c1eaadcac8dce3f74edb9a1e243701f31cb04
     }
 
     /**
@@ -249,6 +269,7 @@ public class Client {
             return;
         }
         httpClient.shutdown();
+        statsControl.shutdown();
         if (authProvider != null) {
             authProvider.close();
         }
@@ -309,9 +330,11 @@ public class Client {
         if (kvRequest.isQueryRequest()) {
             QueryRequest qreq = (QueryRequest)kvRequest;
 
+            statsControl.observeQuery(qreq);
+
             /*
              * The following "if" may be true for advanced queries only. For
-             * such qyeries, the "if" will be true (i.e., the QueryRequest will
+             * such queries, the "if" will be true (i.e., the QueryRequest will
              * be bound with a QueryDriver) if and only if this is not the 1st
              * execute() call for this query. In this case we just return a new,
              * empty QueryResult. Actual computation of a result batch will take
@@ -395,6 +418,7 @@ public class Client {
 
         final long startTime = System.currentTimeMillis();
         kvRequest.setStartTimeMs(startTime);
+        final String requestClass = kvRequest.getClass().getSimpleName();
 
         do {
             long thisTime = System.currentTimeMillis();
@@ -465,6 +489,7 @@ public class Client {
             ResponseHandler responseHandler = null;
 
             ByteBuf buffer = null;
+            long networkLatency;
             try {
                 /*
                  * NOTE: the ResponseHandler will release the Channel
@@ -515,9 +540,10 @@ public class Client {
                                                false /* Don't validate hdrs */);
                 HttpHeaders headers = request.headers();
                 addCommonHeaders(headers);
+                int contentLength = buffer.readableBytes();
                 headers.add(HttpHeaderNames.HOST, host)
                     .add(REQUEST_ID_HEADER, requestId)
-                    .setInt(CONTENT_LENGTH, buffer.readableBytes());
+                    .setInt(CONTENT_LENGTH, contentLength);
 
                 /*
                  * If the request doesn't set an explicit compartment, use
@@ -529,10 +555,10 @@ public class Client {
                 }
                 authProvider.setRequiredHeaders(authString, kvRequest, headers);
 
-                final String requestClass = kvRequest.getClass().getSimpleName();
                 if (isLoggable(logger, Level.FINE)) {
                     logTrace(logger, "Request: " + requestClass);
                 }
+                networkLatency = System.currentTimeMillis();
                 httpClient.runRequest(request, responseHandler, channel);
 
                 boolean isTimeout =
@@ -547,9 +573,12 @@ public class Client {
                              responseHandler.getStatus());
                 }
 
+                ByteBuf wireContent = responseHandler.getContent();
                 Result res = processResponse(responseHandler.getStatus(),
-                                       responseHandler.getContent(),
+                                       wireContent,
                                        kvRequest);
+                int resSize = wireContent.readerIndex();
+                networkLatency = System.currentTimeMillis() - networkLatency;
 
                 if (serialVersion < 3) {
                     /* so we can emit a one-time message if the app */
@@ -585,6 +614,10 @@ public class Client {
 
                 /* copy retry stats to Result on successful operation */
                 res.setRetryStats(kvRequest.getRetryStats());
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+
+                statsControl.observe(kvRequest, Math.toIntExact(networkLatency),
+                    contentLength, resSize);
 
                 return res;
 
@@ -599,6 +632,8 @@ public class Client {
                     exception = rae;
                     continue;
                 }
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+                statsControl.observeError(kvRequest);
                 logInfo(logger, "Unexpected authentication exception: " +
                         rae);
                 throw new NoSQLException("Unexpected exception: " +
@@ -670,10 +705,14 @@ public class Client {
                 }
                 throw upe;
             } catch (NoSQLException nse) {
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+                statsControl.observeError(kvRequest);
                 logFine(logger, "Client execute NoSQLException: " +
                         nse.getMessage());
                 throw nse; /* pass through */
             } catch (RuntimeException e) {
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+                statsControl.observeError(kvRequest);
                 logFine(logger, "Client execute runtime exception: " +
                         e.getMessage());
                 throw e;
@@ -696,12 +735,16 @@ public class Client {
 
                 continue;
             } catch (InterruptedException ie) {
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+                statsControl.observeError(kvRequest);
                 logInfo(logger, "Client interrupted exception: " +
                         ie.getMessage());
                 /* this exception shouldn't retry -- direct throw */
                 throw new NoSQLException("Request interrupted: " +
                                          ie.getMessage());
             } catch (ExecutionException ee) {
+                kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+                statsControl.observeError(kvRequest);
                 logInfo(logger, "Unable to execute request: " +
                         ee.getCause().getMessage());
                 /* is there a better exception? */
@@ -739,6 +782,8 @@ public class Client {
             }
         } while (! timeoutRequest(startTime, timeoutMs, exception));
 
+        kvRequest.setRateLimitDelayedMs(rateDelayedMs);
+        statsControl.observeError(kvRequest);
         throw new RequestTimeoutException(timeoutMs,
             "Request timed out after " + kvRequest.getNumRetries() +
             (kvRequest.getNumRetries() == 1 ? " retry. " : " retries. ") +
@@ -907,7 +952,7 @@ public class Client {
      * Processes the httpResponse object converting it into a suitable
      * return value.
      *
-     * @param httpResponse the response from the service
+     * @param content the response from the service
      *
      * @return the programmatic response object
      */
@@ -979,7 +1024,7 @@ public class Client {
      *
      * @param status the http response code it must not be OK
      *
-     * @param in the input stream representing the failure response
+     * @param payload the payload representing the failure response
      */
     private void processNotOKResponse(HttpResponseStatus status,
                                       ByteBuf payload) {
@@ -1186,6 +1231,13 @@ public class Client {
     }
 
     /**
+     * Returns the statistics control object.
+     */
+    StatsControl getStatsControl() {
+        return statsControl;
+    }
+
+    /**
      * @hidden
      *
      * Try to decrement the serial protocol version.
@@ -1208,7 +1260,7 @@ public class Client {
         return serialVersion;
     }
 
-    /*
+    /**
      * @hidden
      * for internal use
      */