+++ driver/src/main/java/oracle/nosql/driver/ops/MultiDeleteRequest.java
@@ -225,6 +225,11 @@ public class MultiDeleteRequest extends DurableRequest {
         return factory.createMultiDeleteDeserializer();
     }
 
+    @Override
+    public String getTypeName() {
+        return "MultiDelete";
+    }
+
     /**
      * @hidden
      */