+++ driver/src/main/java/oracle/nosql/driver/ops/PrepareRequest.java
@@ -156,6 +156,11 @@ public class PrepareRequest extends Request {
         return factory.createPrepareDeserializer();
     }
 
+    @Override
+    public String getTypeName() {
+        return "Prepare";
+    }
+
     /**
      * @hidden
      */