+++ driver/src/main/java/oracle/nosql/driver/ops/PutRequest.java
@@ -433,6 +433,11 @@ public class PutRequest extends WriteRequest {
         return factory.createPutDeserializer();
     }
 
+    @Override
+    public String getTypeName() {
+        return "Put";
+    }
+
     /**
      * @hidden
      *