+++ driver/src/main/java/oracle/nosql/driver/ops/WriteMultipleRequest.java
@@ -192,6 +192,11 @@ public class WriteMultipleRequest extends DurableRequest {
         return factory.createWriteMultipleDeserializer();
     }
 
+    @Override
+    public String getTypeName() {
+        return "WriteMultiple";
+    }
+
     /**
      * @hidden
      */